requirements.txt
asgiref==3.6.0
boto3==1.26.78
cryptography==39.0.0
Django==3.2.16
django-apscheduler==0.6.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
openai==0.27.2
Pillow==9.4.0
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
gunicorn==20.1.0
typing_extensions==4.4.0
matplotlib==3.3.0
mtcnn==0.1.1
numpy==1.21.6
opencv-python==4.3.0.36
scikit-learn==0.23.1
torch==1.8.0
torchvision===0.9.0
tqdm==4.65.0
tensorflow==2.11.0

